# backend/src/main/resources/openapi.yaml
openapi: 3.1.1
info:
  title: ちいぴよ API
  description: 子供の写真・動画管理及び育児記録管理アプリケーションのAPI定義
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
security:
  - BearerAuth: []
...
paths:
...
  /care-records:
    get:
      tags:
        - careRecordManagement
      summary: 育児記録一覧を取得
      operationId: getCareRecords
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/RequiredStartDate'
        - $ref: '#/components/parameters/RequiredEndDate'
      responses:
        '200':
          description: 育児記録一覧
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/CareRecordListResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    post:
      tags:
        - careRecordManagement
      summary: 育児記録を登録
      operationId: createCareRecord
      x-codegen-request-body-name: CareRecordData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        description: 登録する育児記録情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CareRecordRequestDto'
        required: true
      responses:
        '201':
          description: 記録登録に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /care-records/{id}:
    patch:
      tags:
        - careRecordManagement
      summary: 育児記録を更新
      operationId: updateCareRecord
      x-codegen-request-body-name: CareRecordData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 更新する育児記録情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/CareRecordRequestDto'
        required: true
      responses:
        '204':
          description: 記録更新に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - careRecordManagement
      summary: 育児記録を削除
      operationId: deleteCareRecord
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: 育児記録の削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /growth-records:
    get:
      tags:
        - growthRecordManagement
      summary: 身長・体重記録一覧を取得
      operationId: getGrowthRecords
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/RequiredStartDate'
        - $ref: '#/components/parameters/RequiredEndDate'
      responses:
        '200':
          description: 身長・体重記録一覧
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GrowthRecordResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    post:
      tags:
        - growthRecordManagement
      summary: 身長・体重記録を登録
      operationId: createGrowthRecord
      x-codegen-request-body-name: GrowthRecordData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
      requestBody:
        description: 登録する身長・体重記録情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/GrowthRecordRequestDto'
        required: true
      responses:
        '201':
          description: 身長・体重記録の登録に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
  /growth-records/{id}:
    put:
      tags:
        - growthRecordManagement
      summary: 身長・体重記録を更新
      operationId: updateGrowthRecord
      x-codegen-request-body-name: GrowthRecordData
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        description: 更新する身長・体重記録情報
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/GrowthRecordRequestDto'
        required: true
      responses:
        '204':
          description: 身長・体重記録の更新に成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
    delete:
      tags:
        - growthRecordManagement
      summary: 身長・体重記録を削除
      operationId: deleteGrowthRecord
      parameters:
        - $ref: '#/components/parameters/XRequestedWith'
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: 身長・体重記録の削除に成功
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        default:
          $ref: '#/components/responses/DefaultErrors'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognitoで取得したJWTトークンを指定
  parameters:
    XRequestedWith:
      name: X-Requested-With
      in: header
      required: true
      schema:
        type: string
        default: XMLHttpRequest
    ResourceId:
      name: id
      in: path
      description: リソースの一意な識別子
      required: true
      schema:
        type: integer
        format: int64
...
    RequiredStartDate:
      name: startDate
      in: query
      description: 検索開始日（必須）
      required: true
      schema:
        type: string
        format: date
    RequiredEndDate:
      name: endDate
      in: query
      description: 検索終了日（必須）
      required: true
      schema:
        type: string
        format: date
...
  schemas:
    ErrorResponseDto:
      type: object
      description: エラーレスポンス
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: BAD_REQUEST
        message:
          type: string
          description: エラーメッセージ
          example: リクエストが不正です
...
    MealDetailDto:
      type: object
      description: 食事記録の詳細
      required:
        - note
      properties:
        note:
          type: string
          description: 食事内容・メモ
          example: おかゆと野菜ペースト
    MilkDetailDto:
      type: object
      description: ミルク記録の詳細
      required:
        - amountMl
        - note
      properties:
        amountMl:
          type: integer
          description: 量（ml）
          example: 200
        note:
          type: string
          description: メモ・備考
          example: よく飲んだ
    DiaperDetailDto:
      type: object
      description: 排泄記録の詳細
      required:
        - diaperType
        - note
      properties:
        diaperType:
          type: string
          description: 排泄種別
          enum:
            - WET
            - DIRTY
          example: WET
        note:
          type: string
          description: メモ・備考
    HealthDetailDto:
      type: object
      description: 体調記録の詳細
      required:
        - temperature
        - note
      properties:
        temperature:
          type:
            - number
            - 'null'
          format: double
          description: 体温
          example: 36.5
        note:
          type: string
          description: メモ・備考
          example: 元気
    CareRecordResponseDto:
      type: object
      description: 育児記録情報
      required:
        - id
        - recordedBy
        - recordType
        - recordedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 育児記録ID
          example: 1
        recordedBy:
          type: integer
          format: int64
          description: 記録者のユーザーID
          example: 1
        recordType:
          type: string
          description: 記録種別
          enum:
            - MEAL
            - MILK
            - DIAPER
            - HEALTH
          example: MILK
        recordedAt:
          type: string
          format: date-time
          description: 記録日時
          example: '2026-01-01T08:00:00Z'
        mealDetail:
          anyOf:
            - $ref: '#/components/schemas/MealDetailDto'
            - type: 'null'
        milkDetail:
          anyOf:
            - $ref: '#/components/schemas/MilkDetailDto'
            - type: 'null'
        diaperDetail:
          anyOf:
            - $ref: '#/components/schemas/DiaperDetailDto'
            - type: 'null'
        healthDetail:
          anyOf:
            - $ref: '#/components/schemas/HealthDetailDto'
            - type: 'null'
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2026-01-01T00:00:00Z'
    CareRecordListResponseDto:
      type: object
      description: 育児記録一覧レスポンス
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CareRecordResponseDto'
    CareRecordRequestDto:
      type: object
      description: 育児記録作成・更新リクエスト
      required:
        - recordType
        - recordedAt
      properties:
        recordType:
          type: string
          description: 記録種別
          enum:
            - MEAL
            - MILK
            - DIAPER
            - HEALTH
          example: MILK
        recordedAt:
          type: string
          format: date-time
          description: 記録日時
          example: '2026-01-01T08:00:00Z'
        mealDetail:
          anyOf:
            - $ref: '#/components/schemas/MealDetailDto'
            - type: 'null'
        milkDetail:
          anyOf:
            - $ref: '#/components/schemas/MilkDetailDto'
            - type: 'null'
        diaperDetail:
          anyOf:
            - $ref: '#/components/schemas/DiaperDetailDto'
            - type: 'null'
        healthDetail:
          anyOf:
            - $ref: '#/components/schemas/HealthDetailDto'
            - type: 'null'
    GrowthRecordResponseDto:
      type: object
      description: 身長・体重記録情報
      required:
        - id
        - measurementDate
        - note
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: 記録ID
          example: 1
        measurementDate:
          type: string
          format: date
          description: 測定日
          example: '2026-01-01'
        height:
          type: number
          format: double
          description: 身長（cm）
          example: 75.5
        weight:
          type: number
          format: double
          description: 体重（kg）
          example: 9.5
        note:
          type: string
          description: メモ
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: '2026-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: '2026-01-01T00:00:00Z'
    GrowthRecordRequestDto:
      type: object
      description: 身長・体重記録作成・更新リクエスト
      required:
        - measurementDate
        - note
      properties:
        measurementDate:
          type: string
          format: date
          description: 測定日
          example: '2026-01-01'
        height:
          type:
            - number
            - 'null'
          format: double
          description: 身長（cm）
          example: 75.5
        weight:
          type:
            - number
            - 'null'
          format: double
          description: 体重（kg）
          example: 9.5
        note:
          type: string
          description: メモ
...
  responses:
    DefaultErrors:
      description: 予期しないエラーが発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Unauthorized:
      description: 認証エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    NotFound:
      description: 指定されたリソースが見つからない
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    InternalServerError:
      description: サーバー内部でエラー発生
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'
    Forbidden:
      description: アクセス権限エラー
      content:
        application/json;charset=UTF-8:
          schema:
            $ref: '#/components/schemas/ErrorResponseDto'